server/requirements.txt
flask
psycopg[binary,pool]
python-dotenv